+++ doc/release-notes.md
@@ -4,4 +4,7 @@ Bitcoin ABC version 0.33.4 is now available from:
 
   <https://download.bitcoinabc.org/0.33.4/>
 
-This is a maintenance release with no user-visible change.
+This release includes the following features and fixes:
+ - A new Chronik endpoint `/script/batch/summary` has been added to fetch some
+   limited information from several scripts in a single request (number of txs,
+   utxos and last transaction).